BorsNG.AuthController (Bors-NG v0.1.0)

Routing glue to allow authenticating with GitHub oAuth. BorsNG.Router ensures that users have to go through here. When a user is authenticated, the following items are added to the session

:current_user the ID of the authenticated user in BorsNG.User. This controller will create a user if none exists.

:github_access_token the token to use when accessing the GitHub API.

Functions

callback(conn, map)

This action, which is reached at /auth/:provider/callback, is the the callback URL that the OAuth2 provider will redirect the user back to with a code.

Once here, we will request a permanent access token, allowing us to act on the user's behalf through the GitHub REST API. We also add a row to our database, if it does not already exist.

index(conn, map)

This action is reached via /auth/:provider and redirects to the OAuth2 provider based on the chosen strategy.

logout(conn, params)

socket_token(conn, params)

This action converts a cookie to a socket token. Socket tokens are usable for one hour before they must be reset.