Bors-NG v0.1.0 BorsNG.Database.Attempt

The database-level representation of a "attempt".

When a patch is tried, it gets merged with master individually and it's CI result is reported, but it is not pushed to master.

Functions

all(state)

all_for_patch(patch_id)

all_for_patch(patch_id, state)

all_for_project(project_id, state)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

changeset_state(struct, params \\ %{})

Builds a changeset based on the struct and params.

get_by_commit(project_id, commit, state)

get_next_poll_unix_sec(attempt, project)

new(patch, arguments)

next_poll_is_past(attempt, project)

next_poll_is_past(attempt, project, now_utc_sec)

timeout_is_past(attempt)