Bors-NG v0.1.0 BorsNG.Database.Patch

Corresponds to a pull request in GitHub.

A closed patch may not be r+'ed, nor can a patch associated with a completed batch be r+'ed again, though a patch may be merged and r+'ed at the same time.

Types

t()

t() :: %BorsNG.Database.Patch{
  __meta__: term(),
  author: term(),
  author_id: term(),
  body: term(),
  commit: term(),
  id: term(),
  inserted_at: term(),
  into_branch: term(),
  open: term(),
  pr_xref: term(),
  priority: term(),
  project: term(),
  project_id: term(),
  title: term(),
  updated_at: term()
}

Functions

all(atom)

all_for_batch(batch_id)

all_for_project(project_id, atom)

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

ci_skip?(patch)

ci_skip?(%BorsNG.Database.Patch{
  __meta__: term(),
  author: term(),
  author_id: term(),
  body: term(),
  commit: term(),
  id: term(),
  inserted_at: term(),
  into_branch: term(),
  open: term(),
  pr_xref: term(),
  priority: term(),
  project: term(),
  project_id: term(),
  title: term(),
  updated_at: term()
}) :: boolean()

dups_in_batches()