Bors-NG v0.1.0 BorsNG.Database.Project

Corresponds to a repo in GitHub, as opposed to a repo in Ecto.

This also corresponds to a queue of batches.

Types

t()

t() :: %BorsNG.Database.Project{
  __meta__: term(),
  auto_member_required_perm: term(),
  auto_reviewer_required_perm: term(),
  batch_delay_sec: term(),
  batch_poll_period_sec: term(),
  batch_timeout_sec: term(),
  id: term(),
  inserted_at: term(),
  installation: term(),
  installation_id: term(),
  members: term(),
  name: term(),
  repo_xref: term(),
  staging_branch: term(),
  trying_branch: term(),
  updated_at: term(),
  users: term()
}

Functions

active()

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

changeset_branches(struct, params \\ %{})

changeset_member_settings(struct, params \\ %{})

changeset_reviewer_settings(struct, params \\ %{})

installation_connection(repo_xref, repo)

installation_project_connection(project_id, repo)

orphans()

ping!(project_id)

After modifying the underlying model, call this to notify the UI.