BorsNG.GitHub (Bors-NG v0.1.0)

Wrappers around the GitHub REST API.

Types

tcollaborator()

Specs

tcollaborator() :: %{user: tuser(), perms: tuser_repo_perms()}

tcommitter()

Specs

tcommitter() :: %{name: bitstring(), email: bitstring()}

tconn()

Specs

tconn() :: {ttoken(), number()} | {ttoken(), number()}

A repository connection; it packages a repository with the permissions to access it.

tpr()

Specs

tpr() :: BorsNG.GitHub.Pr.t()

trepo()

Specs

trepo() :: BorsNG.GitHub.Repo.t()

trepo_perm()

Specs

trepo_perm() :: :admin | :push | :pull

tstatus()

Specs

tstatus() :: :ok | :running | :error

ttoken()

Specs

ttoken() :: {:installation, number()} | {:raw, binary()}

An authentication token; this may be a raw token (as on oAuth) or an installation xref (in which case the server will look it up).

tuser()

Specs

tuser() :: BorsNG.GitHub.User.t()

tuser_repo_perms()

Specs

tuser_repo_perms() :: %{admin: boolean(), push: boolean(), pull: boolean()}

Functions

belongs_to_team?(repo_conn, org, team_slug, username)

Specs

belongs_to_team?(tconn(), String.t(), String.t(), String.t()) :: boolean()

create_commit(repo_conn, info)

Specs

create_commit(tconn(), %{
  tree: bitstring(),
  parents: [bitstring()],
  commit_message: bitstring(),
  committer: tcommitter() | nil
}) :: {:ok, binary()} | {:error, term(), term()}

create_commit!(repo_conn, info)

Specs

create_commit!(tconn(), %{
  tree: bitstring(),
  parents: [bitstring()],
  commit_message: bitstring(),
  committer: tcommitter() | nil
}) :: binary()

delete_branch!(repo_conn, branch)

Specs

delete_branch!(tconn(), binary()) :: :ok

force_push!(repo_conn, sha, to)

Specs

force_push!(tconn(), binary(), binary()) :: binary()

get_app!()

Specs

get_app!() :: String.t()

get_branch!(repo_conn, from)

Specs

get_branch!(tconn(), binary()) :: %{commit: bitstring(), tree: bitstring()}

get_collaborators_by_repo(repo_conn)

Specs

get_collaborators_by_repo(tconn()) :: {:ok, [tcollaborator()]} | :error

get_commit_reviews!(repo_conn, issue_xref, sha)

Specs

get_commit_reviews!(tconn(), integer() | bitstring(), binary()) :: map()

get_commit_status!(repo_conn, sha)

Specs

get_commit_status!(tconn(), binary()) :: %{required(binary()) => tstatus()}

get_file!(repo_conn, branch, path)

Specs

get_file!(tconn(), binary(), binary()) :: binary() | nil

get_installation_list!()

Specs

get_installation_list!() :: [integer()]

get_installation_repos!(token)

Specs

get_installation_repos!(ttoken()) :: [trepo()]

get_labels!(repo_conn, issue_xref)

Specs

get_labels!(tconn(), integer() | bitstring()) :: [bitstring()]

get_open_prs!(repo_conn)

Specs

get_open_prs!(tconn()) :: [tpr()]

get_open_prs_with_base!(repo_conn, base)

Specs

get_open_prs_with_base!(tconn(), binary()) :: [tpr()]

get_pr(repo_conn, pr_xref)

Specs

get_pr(tconn(), integer() | bitstring()) ::
  {:ok, BorsNG.GitHub.Pr.t()} | {:error, term()}

get_pr!(repo_conn, pr_xref)

Specs

get_pr!(tconn(), integer() | bitstring()) :: BorsNG.GitHub.Pr.t()

get_pr_commits(repo_conn, pr_xref)

Specs

get_pr_commits(tconn(), integer() | bitstring()) ::
  {:ok, [BorsNG.GitHub.Commit.t()]} | {:error, term()}

get_pr_commits!(repo_conn, pr_xref)

Specs

get_pr_commits!(tconn(), integer() | bitstring()) :: [BorsNG.GitHub.Commit.t()]

get_pr_files(repo_conn, pr_xref)

Specs

get_pr_files(tconn(), integer()) ::
  {:ok, [BorsNG.GitHub.File.t()]} | {:error, term()}

get_pr_files!(repo_conn, pr_xref)

Specs

get_pr_files!(tconn(), integer()) :: [BorsNG.GitHub.File.t()]

get_reviews!(repo_conn, issue_xref)

Specs

get_reviews!(tconn(), integer() | bitstring()) :: map()

get_user_by_login!(token, login)

Specs

get_user_by_login!(ttoken(), binary()) :: {:ok, tuser()} | :error | nil

map_changed_status(check_name)

Specs

map_changed_status(binary()) :: binary()

map_check_to_status(conclusion)

Specs

map_check_to_status(binary()) :: tstatus()

map_state_to_status(state)

Specs

map_state_to_status(binary()) :: tstatus()

map_status_to_state(state)

Specs

map_status_to_state(tstatus()) :: binary()

merge_branch!(repo_conn, info)

Specs

merge_branch!(tconn(), %{
  from: bitstring(),
  to: bitstring(),
  commit_message: bitstring()
}) :: %{commit: binary(), tree: binary()} | :conflict

post_comment!(repo_conn, number, body)

Specs

post_comment!(tconn(), number(), binary()) :: :ok

post_commit_status!(repo_conn, arg)

Specs

post_commit_status!(tconn(), {binary(), tstatus(), binary(), binary()}) :: :ok

push(repo_conn, sha, to)

Specs

push(tconn(), binary(), binary()) ::
  {:ok, binary()} | {:error, term(), term(), term()}

push!(repo_conn, sha, to)

Specs

push!(tconn(), binary(), binary()) :: binary()

synthesize_commit!(repo_conn, info)

Specs

synthesize_commit!(tconn(), %{
  branch: bitstring(),
  tree: bitstring(),
  parents: [bitstring()],
  commit_message: bitstring(),
  committer: tcommitter() | nil
}) :: binary()

update_pr(repo_conn, pr)

Specs

update_pr(tconn(), BorsNG.GitHub.Pr.t()) ::
  {:ok, BorsNG.GitHub.Pr.t()} | {:error, term()}

update_pr!(repo_conn, pr)

Specs

update_pr!(tconn(), BorsNG.GitHub.Pr.t()) :: BorsNG.GitHub.Pr.t()

update_pr_base(repo_conn, pr)

Specs

update_pr_base(tconn(), BorsNG.GitHub.Pr.t()) ::
  {:ok, BorsNG.GitHub.Pr.t()} | {:error, term()}

update_pr_base!(repo_conn, pr)

Specs

update_pr_base!(tconn(), BorsNG.GitHub.Pr.t()) :: BorsNG.GitHub.Pr.t()