BorsNG.GitHub.FriendlyMock (Bors-NG v0.1.0)

Helper functions for ServerMock for common operations without having to modify state by hand.

Tries to lookup values instead of requiring full %{} associative arrays.

Assumes a single GitHub instance with a single repository and single user.

Does everything through webhook notifications. Does not use Database.Repo.insert directly! (One exception: adding a reviewer, which is normally done through Bors' web UI.)

For example, I can run iex -S mix phx.server and do this:

iex> alias BorsNG.GitHub.FriendlyMock
iex> FriendlyMock.init_state
iex> FriendlyMock.make_admin
iex> pr_num = FriendlyMock.add_pr "first"
iex> FriendlyMock.add_reviewer
iex> FriendlyMock.ci_status("SHA-1", "ci", :running)
iex> FriendlyMock.add_comment(pr_num, "bors ping")
iex> FriendlyMock.add_comment(pr_num, "bors r+")
iex> FriendlyMock.get_state
  %{
    {:installation, 91} => %{
          repos: [
            %BorsNG.GitHub.Repo{
              id: 14,
              name: "test/repo",
              owner: %{avatar_url: "", id: 7, login: "tester", type: :user},
              private: false
            }
          ]
    },
    {{:installation, 91}, 14} => %{
          branches: %{"first" => "SHA-1", "master" => "ini"},
          collaborators: %{},
          comments: %{
            1 => [":-1: Rejected by PR status", "bors r+", "pong", "bors ping"]
          },
          commits: %{},
          files: %{
            "SHA-1" => %{
              ".github/bors.toml" => "status = [ "ci" ]

pr_status = [ "ci" ] prerun_timeout_sec = 5 delete_merged_branches = true "

            },
            "master" => %{
              ".github/bors.toml" => "status = [ "ci" ]

pr_status = [ "ci" ] prerun_timeout_sec = 5 delete_merged_branches = true "

            },
            "staging.tmp" => %{
              ".github/bors.toml" => "status = [ "ci" ]

pr_status = [ "ci" ] prerun_timeout_sec = 5 delete_merged_branches = true "

            }
          },
          pr_commits: %{1 => []},
          pulls: %{
            1 => %BorsNG.GitHub.Pr{
              base_ref: "master",
              base_repo_id: 0,
              body: nil,
              head_ref: "first",
              head_repo_id: 0,
              head_sha: "SHA-1",
              merged: false,
              number: 1,
              state: :open,
              title: "first",
              user: %{"avatar_url" => "", "id" => 7, "login" => "tester"}
            }
          },
          statuses: %{"SHA-1" => %{"ci" => :running}}
    }
  }

Now try

iex> FriendlyMock.ci_status("SHA-1", "ci", :ok)

instead of :running above.

Functions

add_comment(pr_num, body, author \\ %{"avatar_url" => "", "id" => 7, "login" => "tester"})

Adds a reviewer comment.

add_commit(pr_num, sha, author)

add_pr(title, body \\ nil)

add_reviewer(repo \\ 14, user \\ %{"avatar_url" => "", "id" => 7, "login" => "tester"})

ci_status(hash, ci_name, status)

Set CI status

comments(repo \\ 14, inst \\ 91)

commits(pr_num, repo \\ 14, inst \\ 91)

full_example()

An example function. Call from iex with

iex> FriendlyMock.full_example

Then modify, recompile() and run again.

get_state()

init_state()

Creates a single installation with a single repo where nothing has happened yet.

make_admin(username \\ %{"avatar_url" => "", "id" => 7, "login" => "tester"}["login"])

pr_to_json(pr)

prs(repo \\ 14, inst \\ 91)

Get open PRs

update_mock(path, fun, repo \\ 14, inst \\ 91)