Bors-NG v0.1.0 BorsNG.GitHub.Server

Provides a real connection to GitHub's REST API. This doesn't currently do rate limiting, but it will.

Types

tcollaborator()

tcollaborator() :: BorsNG.GitHub.tcollaborator()

tconn()

tconn() :: BorsNG.GitHub.tconn()

tpr()

tpr() :: BorsNG.GitHub.tpr()

trepo()

trepo() :: BorsNG.GitHub.trepo()

ttoken()

ttoken() :: BorsNG.GitHub.ttoken()

ttokenreg()

ttokenreg() :: %{optional(number()) => {binary(), number()}}

The token cache.

tuser()

tuser() :: BorsNG.GitHub.tuser()

tuser_repo_perms()

tuser_repo_perms() :: BorsNG.GitHub.tuser_repo_perms()

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_call(atom, repo_conn, pr)

get_collaborators_by_repo_(token, url, append)

get_collaborators_by_repo_(binary(), binary(), [tcollaborator()]) ::
  {:ok, [tcollaborator()]} | {:error, :get_collaborators_by_repo}

get_installation_token!(installation_xref)

get_installation_token!(number()) :: binary()

get_jwt_token()

init(atom)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

args is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

raw_token!(raw, state)

raw_token!(ttoken(), ttokenreg()) :: {{:raw, binary()}, ttokenreg()}

Given an {:installation, installation_xref}, look it up in the token cache. If it's there, and it's still usable, use it. Otherwise, fetch a new one.

start_link()

use_token!(token, state, fun)

use_token!(ttoken(), ttokenreg(), (ttoken() -> term())) ::
  {:reply, term(), ttokenreg()}

Uses a token from the cache, or, if the request fails, retry without using the cached token.