Bors-NG v0.1.0 BorsNG.GitHub.ServerMock

Provides a fake connection to GitHub's REST API.

This is used for unit testing and when running in a "dev" environment, like on a local machine. It's basically just a genserver frontend for a big map; you can put and get its state, and other functions will mutate or read subsets of it.

For example, I can run iex -S mix phx.server and do this:

iex> # Push state to "GitHub"
iex> alias BorsNG.GitHub
iex> alias BorsNG.GitHub.ServerMock
iex> ServerMock.put_state(%{
...>   {:installation, 91} => %{ repos: [
...>     %GitHub.Repo{
...>       id: 14,
...>       name: "test/repo",
...>       owner: %{
...>         id: 6,
...>         login: "bors-fanboi",
...>         avatar_url: "data:image/svg+xml,<svg></svg>",
...>         type: :user
...>       }}
...>   ] },
...>   {{:installation, 91}, 14} => %{
...>     branches: %{},
...>     commits: %{},
...>     comments: %{1 => []},
...>     pulls: %{
...>       1 => %GitHub.Pr{
...>         number: 1,
...>         title: "Test",
...>         body: "Mess",
...>         state: :open,
...>         base_ref: "master",
...>         head_sha: "00000001",
...>         user: %GitHub.User{
...>           id: 6,
...>           login: "bors-fanboi",
...>           avatar_url: "data:image/svg+xml,<svg></svg>"}}},
...>     pr_commits: %{
...>       1 => [
...>         %GitHub.Commit{
...>           sha: "00000001",
...>           author_name: "Bors Fanboi",
...>           author_email: "bors-fanboi@example.com"}
...>       ] },
...>     statuses: %{},
...>     files: %{}}})
iex> GitHub.get_open_prs!({{:installation, 91}, 14})
[
  %BorsNG.GitHub.Pr{
    number: 1,
    title: "Test",
    body: "Mess",
    state: :open,
    base_ref: "master",
    head_sha: "00000001",
    user: %BorsNG.GitHub.User{
      id: 6,
      login: "bors-fanboi",
      avatar_url: "data:image/svg+xml,<svg></svg>"}}]

Types

tbranch()

tbranch() :: bitstring()

tcollaborator()

tcollaborator() :: BorsNG.GitHub.tcollaborator()

tcommit()

tcommit() :: bitstring()

tconn()

tconn() :: BorsNG.GitHub.tconn()

trepo()

trepo() :: BorsNG.GitHub.trepo()

tstate()

tstate() :: %{
  optional(tconn()) => %{
    branches: %{optional(tbranch()) => tcommit()},
    comments: %{optional(integer()) => [bitstring()]},
    commits: %{optional(bitstring()) => tsynthesized()},
    statuses: %{
      optional(tbranch()) => %{
        optional(bitstring()) => :open | :closed | :running
      }
    },
    files: %{optional(tbranch()) => %{optional(bitstring()) => bitstring()}},
    collaborators: [tcollaborator()]
  },
  optional({:installation, number()}) => %{repos: [trepo()]},
  :users => %{optional(bitstring()) => tuser()},
  :merge_conflict => integer()
}

tsynthesized()

tsynthesized() :: %{commit_message: bitstring(), parents: [bitstring()]}

ttoken()

ttoken() :: BorsNG.GitHub.ttoken()

tuser()

tuser() :: BorsNG.GitHub.User.t()

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_call(atom, repo_conn, params, state)

get_state()

init(atom)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

args is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

put_state(state)

start_link()