Bors-NG v0.1.0 BorsNG.ProjectPingChannel

Phoenix channel for notifying users when a repository gets a visible change in its list of PRs.

This channel gets notified when:

the project gets a new patch.

a patch goes from "awaiting review" to "waiting"

a patch goes from "waiting" to "running"

a patch goes from "running" to "awaiting review" (it failed)

a patch goes from "running" to "complete" (it passed)

a patch changes from "open" to "closed", or from "closed" to "open"

Functions

handle_out(topic, msg, socket)

Intercepts outgoing events.

See intercept/1.

Callback implementation for Phoenix.Channel.handle_out/3.

join(arg, message, socket)

Handle channel joins by topic.

To authorize a socket, return {:ok, socket} or {:ok, reply, socket}. To refuse authorization, return {:error, reason}.

Example

def join("room:lobby", payload, socket) do
  if authorized?(payload) do
    {:ok, socket}
  else
    {:error, %{reason: "unauthorized"}}
  end
end

Callback implementation for Phoenix.Channel.join/3.

start_link(triplet)