BorsNG.Router (Bors-NG v0.1.0)

This module maps from URLs to controllers and plugs. It layers on pre-filters, primarily the session, flash, CSRF protection, secure headers, and user authentication part of the session.

Functions

browser_admin(conn, _)

browser_ajax(conn, _)

browser_login(conn, _)

browser_page(conn, _)

browser_session(conn, _)

call(conn, opts)

Callback invoked by Plug on every request.

init(opts)

Callback required by Plug that initializes the router for serving web requests.

webhook(conn, _)