Bors-NG v0.1.0 BorsNG.Worker.Attemptor

An "Attemptor" manages the set of running attempts (that is, "try jobs"). It implements this set of rules:

When a patch is tried, We immediately merge it with master into the trying branch.

The project's CI is occasionally polled, if a attempt is currently running. After polling, the completion logic is run.

If a notification related to the underlying CI is received, the completion logic is run.

When the completion logic is run, the either succeeded or failed.

Functions

cancel(pid, patch_id)

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_cast(arg, project_id)

init(project_id)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

args is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

poll(pid)

start_link(project_id)

status(pid, stat)

tried(pid, patch_id, arguments)