BorsNG.Worker.Batcher (Bors-NG v0.1.0)

A "Batcher" manages the backlog of batches a project has. It implements this set of rules:

When a patch is reviewed ("r+'ed"), it gets added to the project's non-running batch. If no such batch exists, it creates it.

After a short delay, if there is no currently running batch, the project's non-running batch is started.

The project's CI is occasionally polled, if a batch is currently running. After polling, the completion logic is run.

If a notification related to the underlying CI is received, the completion logic is run.

When the completion logic is run, the batch is either bisected (if it failed and there are two or more patches in it), blocked (if it failed and there is only one patch in it), pushed to master (if it passed), or (if there are still CI jobs with no results) it is left alone.

Functions

cancel(pid, patch_id)

cancel_all(pid)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_cast(arg, project_id)

gather_co_authors(batch, patch_links)

get_new_batch(project_id, into_branch, priority, arg4)

init(project_id)

Callback implementation for GenServer.init/1.

poll(pid)

reviewed(pid, patch_id, reviewer)

set_is_single(pid, patch_id, is_single)

set_priority(pid, patch_id, priority)

sort_batches(batches)

start_link(project_id)

status(pid, stat)