Bors-NG v0.1.0 BorsNG.Worker.SyncerInstallation

A background task that pulls a full list of repositories that bors is on. Projects that don't come up get removed, and projects hat don't exist get created.

Functions

add(map1, map2)

plan_synchronize(allow_private, repos, projects)

remove(installation, project)

start_synchronize_all_installations()

start_synchronize_installation(installation)

sync(installation, project)

synchronize_all_installations()

synchronize_installation(installation)

wait_hot_spin_xref(installation_xref)

Wait for synchronization to finish by hot-spinning. Used in test cases.