mix bors.cleanup (Bors-NG v0.1.0)

Delete old batches and patches from the database.

Old patches and batches which are done processing generally are not useful to keep around. Deleting them regularly can keep the rows of your database tables under the small limits of some database providers.

Examples

mix bors.cleanup --months 6

Command line options

--months - the integer number of months to keep data

Functions

repos()