BorsNG.Command (Bors-NG v0.1.0)

Resolve magic comments.

try

The bors comment CLI allows parameters to be passed to try. Assuming the activation phrase is "bors try", you can do things like this:

bors try --layout

And the commit will come out like:

Try #13: --layout

Your build scripts should then inspect the commit message to pull out the commands.

Types

cmd()

Specs

cmd() ::
  {:try, binary()}
  | :try_cancel
  | {:activate_by, binary()}
  | {:set_is_single, integer()}
  | {:set_priority, integer()}
  | :activate
  | :deactivate
  | :delegate
  | {:delegate_to, binary()}
  | {:autocorrect, binary()}
  | :ping
  | :retry

t()

Specs

t() :: %BorsNG.Command{
  comment: binary(),
  commenter: BorsNG.Database.User.t(),
  patch: BorsNG.Database.Patch.t() | nil,
  pr: map() | nil,
  pr_xref: integer(),
  project: BorsNG.Database.Project.t()
}

Functions

delegate_to(c, delegatee)

fetch_patch(c)

Specs

fetch_patch(t()) :: t()

If the Patch is not already in this struct, fetch it. This will not re-sync from GitHub unless it isn't even in the database.

fetch_pr(c)

Specs

fetch_pr(t()) :: t()

If the GitHub PR is not already in this struct, fetch it.

log(c, cmd)

Specs

log(t(), cmd()) :: :ok

parse(comment)

Specs

parse(nil) :: []

parse(binary()) :: [cmd()]

Parse a comment for bors commands.

parse_activation_args(arguments)

parse_activation_args(binary, string)

The username part of an activation-by command is defined like this:

It may start with whitespace

@-signs are stripped

", " is converted to ","

Otherwise, whitespace ends it.

iex> alias BorsNG.Command iex> Command.parse_activation_args("", " this, is, whitespace heavy") "this,is,whitespace" iex> Command.parse_activation_args("", " @this, @has, @ats") "this,has,ats" iex> Command.parse_activation_args("", " trimmed ") "trimmed" iex> Command.parse_activation_args("", "what\never") "what" iex> Command.parse_activation_args("", "") "" iex> Command.parse_activation_args("somebody") [{:activate_by, "somebody"}] iex> Command.parse_activation_args("") [] iex> Command.parse_activation_args(" ") [] iex> Command.parse_activation_args("somebody p=10") [{:set_priority, 10}, {:activate_by, "somebody"}]

parse_cmd(arg1)

parse_delegation_args(arguments)

parse_delegation_args(l, nick, arg3)

The username part of a delegate-to command is defined like this:

It may start with whitespace

@-signs are stripped

", " is converted to ","

Otherwise, whitespace ends it.

It's split on comma.

iex> alias BorsNG.Command iex> Command.parse_delegation_args(" this, is, whitespace heavy") [ {:delegate_to, "this"}, {:delegate_to, "is"}, {:delegate_to, "whitespace"}] iex> Command.parse_delegation_args(" @this, @has, @ats") [{:delegate_to, "this"}, {:delegate_to, "has"}, {:delegate_to, "ats"}] iex> Command.parse_delegation_args(" trimmed ") [{:delegate_to, "trimmed"}] iex> Command.parse_delegation_args("what\never") [{:delegate_to, "what"}] iex> Command.parse_delegation_args("somebody") [{:delegate_to, "somebody"}] iex> Command.parse_delegation_args("") [] iex> Command.parse_delegation_args(" ") []

parse_priority(binary)

parse_single_patch(binary)

permission_denied(c)

regex()

required_permission_level(cmd_list)

required_permission_level_cmd(arg1)

run(c)

Specs

run(t()) :: :ok

Given a populated struct, run everything.

run(c, arg2)

Specs

run(t(), cmd()) :: :ok

trim_and_parse_cmd(arg1)