BorsNG.Database.Attempt (Bors-NG v0.1.0)

The database-level representation of a "attempt".

When a patch is tried, it gets merged with master individually and it's CI result is reported, but it is not pushed to master.

Types

id()

Specs

id() :: pos_integer()

t()

Specs

t() :: %BorsNG.Database.Attempt{
  __meta__: term(),
  arguments: term(),
  commit: term(),
  id: term(),
  inserted_at: term(),
  into_branch: term(),
  last_polled: term(),
  patch: term(),
  patch_id: term(),
  state: term(),
  timeout_at: term(),
  updated_at: term()
}

Functions

all(state)

Specs

all(BorsNG.Database.AttemptState.t() | :incomplete) :: Ecto.Queryable.t()

all_for_patch(patch_id)

Specs

all_for_patch(BorsNG.Database.Patch.id()) :: Ecto.Queryable.t()

all_for_patch(patch_id, state)

Specs

all_for_patch(
  BorsNG.Database.Patch.id(),
  BorsNG.Database.AttemptState.t() | :complete | :incomplete | nil
) :: Ecto.Queryable.t()

all_for_project(project_id, state)

Specs

all_for_project(
  BorsNG.Database.Project.id(),
  BorsNG.Database.AttemptState.t() | :incomplete
) :: Ecto.Queryable.t()

changeset(struct, params \\ %{})

Specs

changeset(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

Builds a changeset based on the struct and params.

changeset_state(struct, params \\ %{})

Specs

changeset_state(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

Builds a changeset based on the struct and params.

get_by_commit(project_id, commit, state)

Specs

get_by_commit(
  BorsNG.Database.Project.id(),
  String.t(),
  BorsNG.Database.AttemptState.t() | :incomplete
) :: Ecto.Queryable.t()

get_next_poll_unix_sec(attempt, project)

Specs

get_next_poll_unix_sec(t(), BorsNG.Database.Project.t()) :: non_neg_integer()

new(patch, arguments)

Specs

new(BorsNG.Database.Patch.t(), String.t()) :: t()

next_poll_is_past(attempt, project)

Specs

next_poll_is_past(t(), BorsNG.Database.Project.t()) :: boolean()

next_poll_is_past(attempt, project, now_utc_sec)

Specs

next_poll_is_past(t(), BorsNG.Database.Project.t(), pos_integer()) :: boolean()

timeout_is_past(attempt)

Specs

timeout_is_past(t()) :: boolean()