BorsNG.Database.Patch (Bors-NG v0.1.0)

Corresponds to a pull request in GitHub.

A closed patch may not be r+'ed, nor can a patch associated with a completed batch be r+'ed again, though a patch may be merged and r+'ed at the same time.

Types

id()

Specs

id() :: pos_integer()

t()

Specs

t() :: %BorsNG.Database.Patch{
  __meta__: term(),
  author: term(),
  author_id: term(),
  body: term(),
  commit: term(),
  id: term(),
  inserted_at: term(),
  into_branch: term(),
  is_single: term(),
  open: term(),
  pr_xref: term(),
  priority: term(),
  project: term(),
  project_id: term(),
  title: term(),
  updated_at: term()
}

Functions

all(atom)

Specs

all(:awaiting_review) :: Ecto.Queryable.t()

all_for_batch(batch_id)

Specs

all_for_batch(BorsNG.Database.Batch.id()) :: Ecto.Queryable.t()

all_for_project(project_id, atom)

Specs

all_for_project(BorsNG.Database.Project.id(), :open | :awaiting_review) ::
  Ecto.Queryable.t()

changeset(struct, params \\ %{})

Specs

changeset(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

Builds a changeset based on the struct and params.

ci_skip?(patch)

Specs

ci_skip?(t()) :: boolean()

dups_in_batches()

Specs

dups_in_batches() :: Ecto.Queryable.t()