BorsNG.Database.Project (Bors-NG v0.1.0)

Corresponds to a repo in GitHub, as opposed to a repo in Ecto.

This also corresponds to a queue of batches.

Types

id()

Specs

id() :: pos_integer()

t()

Specs

t() :: %BorsNG.Database.Project{
  __meta__: term(),
  auto_member_required_perm: term(),
  auto_reviewer_required_perm: term(),
  batch_delay_sec: term(),
  batch_poll_period_sec: term(),
  batch_timeout_sec: term(),
  id: term(),
  inserted_at: term(),
  installation: term(),
  installation_id: term(),
  members: term(),
  name: term(),
  repo_xref: term(),
  staging_branch: term(),
  trying_branch: term(),
  updated_at: term(),
  users: term()
}

Functions

active()

Specs

active() :: Ecto.Queryable.t()

changeset(struct, params \\ %{})

Specs

changeset(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

Builds a changeset based on the struct and params.

changeset_branches(struct, params \\ %{})

Specs

changeset_branches(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

changeset_member_settings(struct, params \\ %{})

Specs

changeset_member_settings(t() | Ecto.Changeset.t(), map()) :: Ecto.Changeset.t()

changeset_reviewer_settings(struct, params \\ %{})

Specs

changeset_reviewer_settings(t() | Ecto.Changeset.t(), map()) ::
  Ecto.Changeset.t()

installation_connection(repo_xref, repo)

Specs

installation_connection(repo_xref, module()) ::
  {{:installation, BorsNG.Database.Installation.xref()}, repo_xref}
when repo_xref: integer()

installation_project_connection(project_id, repo)

Specs

installation_project_connection(id(), module()) ::
  {{:installation, BorsNG.Database.Installation.xref()}, repo_xref :: integer()}

orphans()

Specs

orphans() :: Ecto.Queryable.t()

ping!(project_id)

Specs

ping!(id() | binary()) :: :ok | no_return()

After modifying the underlying model, call this to notify the UI.