BorsNG.GitHub.Server (Bors-NG v0.1.0)

Provides a real connection to GitHub's REST API. This doesn't currently do rate limiting, but it will.

Types

tcollaborator()

Specs

tcollaborator() :: BorsNG.GitHub.tcollaborator()

tconn()

Specs

tconn() :: BorsNG.GitHub.tconn()

tpr()

Specs

tpr() :: BorsNG.GitHub.tpr()

trepo()

Specs

trepo() :: BorsNG.GitHub.trepo()

ttoken()

Specs

ttoken() :: BorsNG.GitHub.ttoken()

ttokenreg()

Specs

ttokenreg() :: %{required(number()) => {binary(), number()}}

The token cache.

tuser()

Specs

tuser() :: BorsNG.GitHub.tuser()

tuser_repo_perms()

Specs

tuser_repo_perms() :: BorsNG.GitHub.tuser_repo_perms()

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_call(atom, repo_conn, pr)

get_collaborators_by_repo_(token, url, append)

Specs

get_collaborators_by_repo_(binary(), binary(), [tcollaborator()]) ::
  {:ok, [tcollaborator()]} | {:error, :get_collaborators_by_repo}

get_installation_token!(installation_xref)

Specs

get_installation_token!(number()) :: binary()

get_jwt_token()

init(atom)

Callback implementation for GenServer.init/1.

raw_token!(raw, state)

Specs

raw_token!(ttoken(), ttokenreg()) :: {{:raw, binary()}, ttokenreg()}

Given an {:installation, installation_xref}, look it up in the token cache. If it's there, and it's still usable, use it. Otherwise, fetch a new one.

start_link()

use_token!(token, state, fun)

Specs

use_token!(ttoken(), ttokenreg(), (ttoken() -> term())) ::
  {:reply, term(), ttokenreg()}

Uses a token from the cache, or, if the request fails, retry without using the cached token.