BorsNG.GitHub.ServerMock (Bors-NG v0.1.0)

Provides a fake connection to GitHub's REST API.

This is used for unit testing and when running in a "dev" environment, like on a local machine. It's basically just a genserver frontend for a big map; you can put and get its state, and other functions will mutate or read subsets of it.

For example, I can run iex -S mix phx.server and do this:

iex> # Push state to "GitHub"
iex> alias BorsNG.GitHub
iex> alias BorsNG.GitHub.ServerMock
iex> ServerMock.put_state(%{
...>   {:installation, 91} => %{ repos: [
...>     %GitHub.Repo{
...>       id: 14,
...>       name: "test/repo",
...>       owner: %{
...>         id: 6,
...>         login: "bors-fanboi",
...>         avatar_url: "data:image/svg+xml,<svg></svg>",
...>         type: :user
...>       }}
...>   ] },
...>   {{:installation, 91}, 14} => %{
...>     branches: %{},
...>     commits: %{},
...>     comments: %{1 => []},
...>     pulls: %{
...>       1 => %GitHub.Pr{
...>         number: 1,
...>         title: "Test",
...>         body: "Mess",
...>         state: :open,
...>         base_ref: "master",
...>         head_sha: "00000001",
...>         user: %GitHub.User{
...>           id: 6,
...>           login: "bors-fanboi",
...>           avatar_url: "data:image/svg+xml,<svg></svg>"}}},
...>     pr_commits: %{
...>       1 => [
...>         %GitHub.Commit{
...>           sha: "00000001",
...>           author_name: "Bors Fanboi",
...>           author_email: "bors-fanboi@example.com"}
...>       ] },
...>     statuses: %{},
...>     files: %{}}})
iex> GitHub.get_open_prs!({{:installation, 91}, 14})
[
  %BorsNG.GitHub.Pr{
    number: 1,
    title: "Test",
    body: "Mess",
    state: :open,
    base_ref: "master",
    head_sha: "00000001",
    user: %BorsNG.GitHub.User{
      id: 6,
      login: "bors-fanboi",
      avatar_url: "data:image/svg+xml,<svg></svg>"}}]

Types

tbranch()

Specs

tbranch() :: bitstring()

tcollaborator()

Specs

tcollaborator() :: BorsNG.GitHub.tcollaborator()

tcommit()

Specs

tcommit() :: bitstring()

tconn()

Specs

tconn() :: BorsNG.GitHub.tconn()

trepo()

Specs

trepo() :: BorsNG.GitHub.trepo()

tstate()

Specs

tstate() :: %{
  required(tconn()) => %{
    branches: %{required(tbranch()) => tcommit()},
    comments: %{required(integer()) => [bitstring()]},
    commits: %{required(bitstring()) => tsynthesized()},
    statuses: %{
      required(tbranch()) => %{
        required(bitstring()) => :open | :closed | :running
      }
    },
    files: %{required(tbranch()) => %{required(bitstring()) => bitstring()}},
    collaborators: [tcollaborator()]
  },
  required({:installation, number()}) => %{repos: [trepo()]},
  :users => %{required(bitstring()) => tuser()},
  :merge_conflict => integer()
}

tsynthesized()

Specs

tsynthesized() :: %{commit_message: bitstring(), parents: [bitstring()]}

ttoken()

Specs

ttoken() :: BorsNG.GitHub.ttoken()

tuser()

Specs

tuser() :: BorsNG.GitHub.User.t()

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

do_handle_call(atom, repo_conn, pr, state)

get_state()

init(atom)

Callback implementation for GenServer.init/1.

put_state(state)

start_link()