Bors-NG v0.1.0 BorsNG.ProjectController

Shows a list of repositories, a single repository, and the repository's settings page.

n.b. We call it a project internally, though it corresponds to a GitHub repository. This is to avoid confusing a GitHub repo with an Ecto repo.

Functions

add_member(conn, atom, project, arg4)

add_reviewer(conn, atom, project, arg4)

cancel_all(conn, atom, project, arg4)

confirm_add_reviewer(conn, atom, project, arg4)

index(conn, arg2)

log(conn, atom, project, arg4)

remove_member(conn, atom, project, arg4)

remove_reviewer(conn, atom, project, arg4)

settings(conn, atom, project, arg4)

show(conn, mode, project, params)

synchronize(conn, atom, project, arg4)

update_branches(conn, atom, project, arg4)

update_member_settings(conn, atom, project, arg4)

update_reviewer_settings(conn, atom, project, arg4)