BorsNG.Router.Helpers (Bors-NG v0.1.0)

Module with named helpers generated from BorsNG.Router.

Functions

admin_path(conn_or_endpoint, action)

admin_path(conn_or_endpoint, action, params)

admin_url(conn_or_endpoint, action)

admin_url(conn_or_endpoint, action, params)

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, params)

auth_path(conn_or_endpoint, action, provider, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, params)

auth_url(conn_or_endpoint, action, provider, params)

batch_path(conn_or_endpoint, action, id)

batch_path(conn_or_endpoint, action, id, params)

batch_url(conn_or_endpoint, action, id)

batch_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

project_path(conn_or_endpoint, action)

project_path(conn_or_endpoint, action, params)

project_path(conn_or_endpoint, action, id, params)

project_path(conn_or_endpoint, action, id, login, params)

project_url(conn_or_endpoint, action)

project_url(conn_or_endpoint, action, params)

project_url(conn_or_endpoint, action, id, params)

project_url(conn_or_endpoint, action, id, login, params)

server_path(conn_or_endpoint, action)

server_path(conn_or_endpoint, action, params)

server_url(conn_or_endpoint, action)

server_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

webhook_path(conn_or_endpoint, action, provider)

webhook_path(conn_or_endpoint, action, provider, params)

webhook_url(conn_or_endpoint, action, provider)

webhook_url(conn_or_endpoint, action, provider, params)